Add support for passing the rack environment along to the resource.
The reason for implementing this is primarily because we want to mount
RackDAV::Handler into a Rails application using the Rails router to
split paths.

For example:
mount "/(:domain)" => RackDAV::Handler.new(:resource_class => OurResource)

When requests come in we are then able to pull out :domain in
OurResource rather than having to parse the path again or write a router
resource.